fix errors and warnings in MSVC when compiling as C++
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@9200 d7cf8633-e32d-0410-b094-e92efae38249

+++ prvm_cmds.c
@@ -3877,7 +3877,7 @@ void VM_bitshift (void)
 	int n1, n2;
 	VM_SAFEPARMCOUNT(2, VM_bitshift);
 
-	n1 = (int)fabs((int)PRVM_G_FLOAT(OFS_PARM0));
+	n1 = (int)fabs((float)((int)PRVM_G_FLOAT(OFS_PARM0)));
 	n2 = (int)PRVM_G_FLOAT(OFS_PARM1);
 	if(!n1)
 		PRVM_G_FLOAT(OFS_RETURN) = n1;
